# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_user_service.BlockStruct:
    properties:
      number:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_user_service.CheckUserStruct:
    properties:
      formatJid:
        type: boolean
      number:
        items:
          type: string
        type: array
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_user_service.GetAvatarStruct:
    properties:
      number:
        type: string
      preview:
        type: boolean
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_user_service.PrivacyStruct:
    properties:
      callAdd:
        $ref: '#/definitions/types.PrivacySetting'
      groupAdd:
        $ref: '#/definitions/types.PrivacySetting'
      lastSeen:
        $ref: '#/definitions/types.PrivacySetting'
      online:
        $ref: '#/definitions/types.PrivacySetting'
      profile:
        $ref: '#/definitions/types.PrivacySetting'
      readReceipts:
        $ref: '#/definitions/types.PrivacySetting'
      status:
        $ref: '#/definitions/types.PrivacySetting'
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_user_service.SetProfilePictureStruct:
    properties:
      image:
        type: string
    type: object
...
  types.PrivacySetting:
    enum:
    - ""
    - all
    - contacts
    - contact_allowlist
    - contact_blacklist
    - match_last_seen
    - known
    - none
    - on_standard
    - "off"
    type: string
    x-enum-varnames:
    - PrivacySettingUndefined
    - PrivacySettingAll
    - PrivacySettingContacts
    - PrivacySettingContactAllowlist
    - PrivacySettingContactBlacklist
    - PrivacySettingMatchLastSeen
    - PrivacySettingKnown
    - PrivacySettingNone
    - PrivacySettingOnStandard
    - PrivacySettingOff
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
...
  /user/avatar:
    post:
      consumes:
      - application/json
      description: Get a user's avatar
      parameters:
      - description: Avatar data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_user_service.GetAvatarStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a user's avatar
      tags:
      - User
  /user/block:
    post:
      consumes:
      - application/json
      description: Block a contact
      parameters:
      - description: Block data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_user_service.BlockStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Block a contact
      tags:
      - User
  /user/blocklist:
    get:
      consumes:
      - application/json
      description: Get a user's block list
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a user's block list
      tags:
      - User
  /user/check:
    post:
      consumes:
      - application/json
      description: Check a user
      parameters:
      - description: User data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_user_service.CheckUserStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Check a user
      tags:
      - User
  /user/contacts:
    get:
      consumes:
      - application/json
      description: Get a user's contacts
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a user's contacts
      tags:
      - User
  /user/info:
    post:
      consumes:
      - application/json
      description: Get a user
      parameters:
      - description: User data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_user_service.CheckUserStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a user
      tags:
      - User
  /user/privacy:
    get:
      consumes:
      - application/json
      description: Get a user's privacy settings
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a user's privacy settings
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Set a user's privacy settings
      parameters:
      - description: Privacy data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_user_service.PrivacyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set a user's privacy settings
      tags:
      - User
  /user/profileName:
    post:
      consumes:
      - application/json
      description: Set a user's profile name
      parameters:
      - description: Profile name data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_user_service.SetProfilePictureStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set a user's profile name
      tags:
      - User
  /user/profilePicture:
    post:
      consumes:
      - application/json
      description: Set a user's profile picture
      parameters:
      - description: Profile picture data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_user_service.SetProfilePictureStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set a user's profile picture
      tags:
      - User
  /user/profileStatus:
    post:
      consumes:
      - application/json
      description: Set a user's profile status
      parameters:
      - description: Profile status data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_user_service.SetProfilePictureStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set a user's profile status
      tags:
      - User
  /user/unblock:
    post:
      consumes:
      - application/json
      description: Unblock a contact
      parameters:
      - description: Block data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_user_service.BlockStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Unblock a contact
      tags:
      - User
swagger: "2.0"